FIX: Appliance clean-up and bug fixing.
Some refactoring has been done to keep naming consistent amongst agents.
