+++ bin/.gitignore
@@ -1,2 +1,3 @@
 /agents/
 /ontologies/
+/utility/
