--- src/agents/AppliancesAgent.java
@@ -1,223 +1,0 @@
-package agents;
-
-import jade.core.*;
-import jade.core.behaviours.*;
-import jade.domain.DFService;
-import jade.domain.FIPANames;
-import jade.domain.FIPAAgentManagement.DFAgentDescription;
-import jade.domain.FIPAAgentManagement.FailureException;
-import jade.domain.FIPAAgentManagement.NotUnderstoodException;
-import jade.domain.FIPAAgentManagement.RefuseException;
-import jade.domain.FIPAAgentManagement.ServiceDescription;
-import jade.lang.acl.ACLMessage;
-import jade.lang.acl.MessageTemplate;
-import jade.proto.AchieveREResponder;
-import jade.proto.SubscriptionResponder;
-import jade.proto.SubscriptionResponder.Subscription;
-import jade.content.ContentElement;
-import jade.content.lang.*;
-import jade.content.lang.sl.*;
-import jade.content.onto.*;
-import jade.content.onto.basic.Action;
-
-import java.util.Random;
-
-import ontologies.*;
-import ontologies.Retailer.retailerType;
-import utility.*;
-
-@SuppressWarnings("serial")
-public class AppliancesAgent extends Agent implements SupplierVocabulary{
-	private Appliances appliances = new Appliances();
-	private Codec codec = new SLCodec();
-	private Ontology ontology = SupplierOntology.getInstance();
-	private Random rnd = Utility.newRandom(hashCode());
-	
-	private static final int TICK_TIME = (60000 * 5);
-	
-	void setupAppliances() {
-		appliances.setGenerationRate(rnd.nextInt(10)); //random int between 0-10
-		appliances.setUsageRate(rnd.nextInt(30));    //random int between 0-30
-		appliances.setSupply(rnd.nextInt(2000)); //initial supply random 0-2000
-		
-		System.out.println(appliances.toString());
-		addBehaviour(updateAppliances);
-	}
-	
-	protected void setup() {
-		// Register language and ontology
-		getContentManager().registerLanguage(codec);
-		getContentManager().registerOntology(ontology);
-		
-		// Setup retailer state
-		setupAppliances();
-		
-		// Register in the DF
-		DFRegistry.register(this, APPLIANCE_AGENT);
-		subscriptionHandler();
-		
-		// Run agent
-		process();
-	}
-	
-	// Deregister this agent
-	protected void takeDown() {
-		System.out.println("Shutting down " + this.getName() + ".");
-		try { DFService.deregister(this); } catch (Exception e) { e.printStackTrace(); };
-		
-		try {
-			DFAgentDescription dfd = new DFAgentDescription();
-			ServiceDescription sd = new ServiceDescription();
-			sd.setType(APPLIANCE_AGENT);
-			dfd.addServices(sd);
-			
-			DFAgentDescription[] dfds = DFService.search(this, dfd);
-			
-			if(dfds.length > 0) {
-				System.out.println("\tRemaining appliances:");
-			} else {
-				System.out.println("\tNo remaining appliances.");
-			}
-			
-			for(int i = 0; i < dfds.length; i++) {
-				System.out.print("\t" + dfds[i].getName());
-			}
-		} catch (Exception e) { e.printStackTrace(); }
-	}
-	
-	void process() {
-		System.out.println("Agent " + getLocalName() +  " waiting for requests...");
-		MessageTemplate template = MessageTemplate.and(
-		  		MessageTemplate.MatchProtocol(FIPANames.InteractionProtocol.FIPA_REQUEST),
-		  		MessageTemplate.MatchPerformative(ACLMessage.REQUEST) );
-		
-		// Handle purchase confirmation
-		addBehaviour(new AchieveREResponder(this, template) {
-			@Override
-			protected ACLMessage handleRequest(ACLMessage request) throws NotUnderstoodException, RefuseException {
-				System.out.println("Agent " + getLocalName() + ": REQUEST received from " + 
-						request.getSender().getName() + ". Action is " + request.getContent());
-		
-				System.out.println("Agent " + getLocalName() + ": Agree");
-				ACLMessage agree = request.createReply();
-				agree.setPerformative(ACLMessage.AGREE);
-				return agree;
-			}
-			
-			@Override
-			protected ACLMessage prepareResultNotification(ACLMessage request, ACLMessage response) throws FailureException {
-				ACLMessage inform = request.createReply();
-				inform.setPerformative(ACLMessage.INFORM);
-				request.setLanguage(codec.getName());
-				request.setOntology(ontology.getName());
-				
-				try {
-					ContentElement content = getContentManager().extractContent(request);
-					Exchange action = (Exchange)((Action)content).getAction();
-					
-					switch(action.getType()) {
-					case SupplierVocabulary.BUY:
-						int updatedSupply = appliances.getSupply() - action.getUnits();
-						System.out.println("Agent " + getLocalName() + ": Action successfully performed" + 
-								"\nSuppy held: " + appliances.getSupply() + "\nRemaining supply: " + updatedSupply);
-						appliances.setSupply(updatedSupply);
-						
-						break;
-					case SupplierVocabulary.SELL:
-					}
-					
-				} catch(Exception e) {
-					inform.setPerformative(ACLMessage.FAILURE);
-					e.printStackTrace();
-				}
-				
-				System.out.println(inform.getContent());
-				return inform;
-			}
-		} );
-		
-		template = MessageTemplate.and(
-		  		MessageTemplate.MatchProtocol(FIPANames.InteractionProtocol.FIPA_REQUEST),
-		  		MessageTemplate.MatchPerformative(ACLMessage.QUERY_REF) );
-		
-		// Handles quote requests
-		addBehaviour(new AchieveREResponder(this, template) {
-			@Override
-			protected ACLMessage handleRequest(ACLMessage request) throws NotUnderstoodException, RefuseException {
-				System.out.println("Agent " + getLocalName() + ": QUERY_REF received from " + 
-						request.getSender().getName() + ". Action is " + request.getContent());
-				
-				System.out.println("Agent " + getLocalName() + ": Agree");
-				ACLMessage agree = request.createReply();
-				agree.setPerformative(ACLMessage.AGREE);
-				return agree;
-			}
-			
-			
-			// what would be placed here????
-			@Override
-			protected ACLMessage prepareResultNotification(ACLMessage request, ACLMessage response) throws FailureException {
-				System.out.println("Agent " + getLocalName() + ": Quote generated");
-				ACLMessage inform = request.createReply();
-				inform.setPerformative(ACLMessage.INFORM);
-				inform.setContent("" + retailer.getPricePerUnit());
-				return inform;
-			}
-		} );
-	}
-	
-	TickerBehaviour updateAppliances = new TickerBehaviour(this, rnd.nextInt(TICK_TIME)) {
-		@Override
-		public void onTick() {
-			//Update supply based on generation rate
-			appliances.setSupply(appliances.getSupply() + appliances.getGenerationRate());
-		
-			System.out.println(myAgent.getLocalName() + " updating supply...\n Change = " + 
-					appliances.getGenerationRate() + "\n Supply = " + appliances.getSupply());
-			
-			reset(rnd.nextInt(TICK_TIME));
-		}
-	};
-	
-	// -- Utility Methods -- 
-	void subscriptionHandler() {
-		MessageTemplate mt = MessageTemplate.and(
-			MessageTemplate.MatchPerformative(ACLMessage.QUERY_REF),
-			MessageTemplate.MatchProtocol(FIPANames.InteractionProtocol.FIPA_SUBSCRIBE)
-		);
-		
-		SubscriptionResponder sr = new SubscriptionResponder(this, mt) {
-			ACLMessage notification;
-			Subscription sub;
-			int lastPrice = 0;
-			
-			@Override
-			protected ACLMessage handleSubscription(ACLMessage subscription) throws NotUnderstoodException, RefuseException {
-				super.handleSubscription(subscription);
-				System.out.println("Subscription: \n\t" + 
-						subscription.getSender().getName() + " successfully subscribed to " + myAgent.getName());
-				
-				sub = getSubscription(subscription);
-				notification = subscription.createReply();
-				notification.setPerformative(ACLMessage.INFORM);
-				
-				
-				//will this be when the appliance are getting low on stock??
-				addBehaviour(new CyclicBehaviour() {
-					@Override
-					public void action() {
-						if(lastPrice == 0 || retailer.getPricePerUnit() != lastPrice) {
-							notification.setContent(Integer.toString(retailer.getPricePerUnit()));
-							lastPrice = retailer.getPricePerUnit();
-							sub.notify(notification);
-						}
-					}
-				});
-				
-				return null;
-			}	
-		};
-		
-		addBehaviour(sr);
-	}
-}
