+++ src/ontologies/Appliance.java
@@ -3,16 +3,13 @@ package ontologies;
 import jade.content.*;
 
 @SuppressWarnings("serial")
-
-public class Appliances implements Concept{
+public class Appliance implements Concept {
 	private int generationRate;
 	private int usageRate;
-	private int supply;
 	
-	public Appliances() { 
+	public Appliance() { 
 		generationRate = 0;
 		usageRate = 0;
-		supply = 0;
 	}
 	
 	public int getGenerationRate() {
@@ -30,21 +27,14 @@ public class Appliances implements Concept{
 	public void setUsageRate(int usageRate) {
 		this.usageRate = usageRate;
 	}
-
-	public int getSupply() {
-		return supply;
+	
+	public int getRate() {
+		return generationRate - usageRate;
 	}
 
-	public void setSupply(int supply) {
-		this.supply = supply;
-		if (this.supply < 0) this.supply = 0; // Cannot be less than 0
-	}
-	
-	
 	@Override
 	public String toString() {
-		return "Appliances [generationRate=" + generationRate + ", usageRate=" + usageRate + ", supply=" + supply
-				+ "]";
+		return "Appliances [generationRate=" + generationRate + ", " + 
+				"usageRate=" + usageRate + "]";
 	}
-
 }
